Fix repo .gitignore file matching
Without this patch, the untracked_unignored_files method tries to match
files based on regex when .gitignore generally specifies matches with
globs. This causes an error if a user adds a .gitignore to a module
with an entry that is a glob rather than an exact match. This patch
fixes the problem by using File.fnmatch instead of the string match
method to compare file names.