Refactoring hook.rb
This commit turns hook.rb into a class. Now instead of using the module
method 'hook' with a command of 'activate' or 'deactivate' with args,
you instantiate a Hook and pass it the name of a hook file
('.git/hooks/pre-push') and call the activate or deactivate methods. Where,
previously, the hook file path was hardcoded as the `HOOK_FILE` constant,
it is now configurable. Included in this commit are tests for the Hook
class and the Hook CLI interface. The CLI changes will be introduced in
the next commit.