Refactoring the CLI and adding the Project class
This commit addresses #21 and transitions away from OptionParser and
uses Thor to handle the CLI. Thor has a great interface and allows you
to define your CLI using a simple Rake-like DSL. This commit also adds
the Project class.

This commit re-defines our concept of a project root. Previously the
PROJECT_ROOT constant was set to './modules' which is really just
where the modules live. This commit redefines project root as the
root of the modulesync project. Inside of a project root is the
moduleroot (the root of the module template) and the modules directory.

The Project class is really where the meat of this refactor is. I've
made all of our constants available as configuration options. This
allows you to store your modules outside of the project root by setting
'modules_dir' to a fully qualified path in 'modulesync.yaml'.

The Project class is instantiated with a required root parameter (for
the project root) and an optional settings hash. The CLI class makes use
of this by passing the command line arguments to Project via the
settings parameter. The Project class is the container for most of the
functionality of ModuleSync. The CLI class will instantiate a Project,
and will call the `update` method on each of the projects modules. This
is a much more object oriented approach than we had before. The Project
knows what modules it should manage.

The Project class is responsible for reading the project's configuration
file: `modulesync.yml`. It then takes those settings and merges them
with the command line arguments (giving precedence to the command line
arguments) and then merges that set with the default settings (giving
precedence to the merged set of CLI and config file arguments). This
happens in the Project#config method. The default_settings method is
just a mechanism to isolate our defaults. This is where most of our
settings from the constants ended up.

The Project will lazily instantiate Hook objects when Project#hook is
called. Same goes for Project#moduleroot. I created a ModuleRoot class
because it felt like a separate object in and of itself. The Project
class has a `defaults` method that loads the `config_defaults.yml`
file and a `global_defaults` method that looks for the
`global_defaults_key` inside the `config_defaults.yml`. The
`global_defaults_key` was previously a constant, but is now a project
default and is configurable in `modulesync.yml`.

The Project class has a `module_list` method that reads the
`managed_modules_file` (previously a constant, now a configuration
parameter) which defaults to `managed_modules.yml`. This method also
checks to make sure that the modules file exists and that it's not
empty. I've added a `module_list` attribute writer so the user can
manage the modlue_list using some dynamic source as opposed to a static
file. The `module_list` method also filters through the list and rejects
any modules that don't match the filter defined at the CLI with the `-f`
flag (if one was set).

The `Project#modules` method then calls the `module_list` and
instantiates a Module class for each managed module, passing the module
name, self (so, the Project object), and any options defined in
`managed_modules.yml`. It wasn't explicitly documented, but there was
already support for `managed_modules.yml` to contain a single array of
modules or a hash with the module name as the key and some settings.