Moving Renderer into ModuleFile
This commit removes the Renderer module and creates a new class for this
functionality: ModuleFile. The ModuleFile class represents a file being
managed inside of a module listed in `managed_modules.yml`. It is
initialized with two manadatory arguments, the first being a Module
object (the module to which this file belongs), and the name of the file
being managed. The write method actually writes the file and returns the
name of the destination file.