Adding the Module and ModuleRoot classes
The `Module` class gets instantiated once for each module that exists
in a project. The `Module` class requires two parameters to be passed
during initialization, the name of the module, and the project object
that this module belongs to. There's an optional third parameter that
can be passed to the `Module` class for module-specific options (this is
discussed later, when we get to the module hooks). I would like to
decouple the Module class from the Project class, but I think the best
strategy, for now, is dependency injection.

There are a number of attributes that are called on the `Project` object
that gets passed to the `Module` class, such as the `project_root`,
`defaults`, `global_defaults`, `project_hooks`, and `git_opts`. These
are all bits of information that the `Module` class uses to manage the
module. I initially wanted to pass these as arguments to the
`initialize` method, but then the Project class would need to know a lot
about the Module class, I opted to mangage the dependencies in the other
direction. The `Module` class also calls the `source_files` method on
the `Project` object's `ModuleRoot` instance (via the
`Project#moduleroot` method).

The `Module#module_configs` method reads the module's `.sync.yml` file.
The `Module#files` method loads all of the source files inside the
`moduleroot` and finds the union between the files in the moduleroot,
the files listed in `config_defaults.yml` and the files listed in
`.sync.yml`. The `Module#file_configs` method takes a filename as an
argument and returns the configurations that should be passed to the
template for that file.

The `Module#sync` method then iterates over the array of files returned
by the `Module#files` method, calls `Module#file_configs` for each file,
and instantiates a `ModuleFile` class for each managed file unless it's
configured as 'unmanaged'. If it's configured as a 'delete' file, the
sync method calls the `delete!` method and that deletes the file from
the module. If the file is not marked for deletion, the write method is
called on the `ModuleFile`.

The ModuleRoot class is just a container for the list of source files.
It's primary consumer is the Module class. It is initialized with the
path to the module root (defaults to ./moduleroot) and checks to make
sure that it exists. The `ModuleRoot#source_files` method finds and
returns a list all of the files under the module root.