Moving the Git module into the Git::Repo class
This commit removes the Git module and moves its functionality into the
Git::Repo class. I called it Git::Repo instead of Git because a repo is
a thing and a Git is not. I pulled a lot of functionality out of this as
well. I removed the ability to update a changelog because, per the
discussion on the PR, this functionality is so specific as to be useless
the way it was implemented. I've also removed the OpenStack project's
hooks and the functionality to bump a module's version number.

In place of the functionality that I've removed, I've added the ability
to run a pre_commit and pre_push hook. This is configurable down to the
individual module and can be used to re-implement the functionality I
removed above in a vendor-agnostic way. This also enables a whole host
of use-cases that we haven't predicted.

I've also made the output of the git commands a little more readable. I
also broke up the functionality in the Git module into methods so we can
more easily test this class. The Git module was so tightly coupled and
the logic had so many branches as to make it nearly impossible to extend
of test.

The git class still needs some love. I'd also love to see other VCS
support.

I'm also adding some new tests for the `update` cucumber feature, and
updating the existing ones with the new output from the Git::Repo class.