Removing logic from modulesync.rb
All of the logic previously encapsulated in `modulesync.rb` has been
moved into classes. This module contains some error classes, and
contains a singleton `validate_hooks` method. This method takes two
arguments: an array of hooks and a project root. This method ensures
that the hook files exist and that they are executable. It then returns
a list of valid hooks available under the project root.

This commit also includes the SimpleCov settings for Cucumber and
removes the now unused constants module.