Refactoring the Hook module into a class
This commit moves ModuleSync::Hook from a module to a class. It also
adds the ability to define arbitrary arguments to add to the hook. I've
also broken up the `run` method. Instead of calling a monolithic `run`
method we now call `update` or `hook` methods based on the CLI option.