Removing constants
This PR removes the constants module and moves towards a more composable
configuration model. I've made some changes to the cucumber features so
the tests fail more descriptively. I've also added a Config class that
takes over much of the responsiblity of the Constants module. I've moved
responsiblity for maintaining configuration defaults to the Config class
and removed them from the CLI class. I've also renamed project_root
because it was confusing. Now the project_root refers to the directory
where the modulesync.yml file lives. The previous project_root referred
to the `modules/` directory and so I have very creatively renamed it to
`modules_dir`.