Monkey Patch git ruby gem
Previously modulesync would throw an error when a file had a line ending warning
emitted during a diff.  This is common on cross platform modules.  This commit
monkey patches the git ruby gem to handle extra text before the actual diff
contents.  A monkey patch was used as the git ruby gem does not seem to be
actively maintained, and it will some time before this issue is resolved and a
new gem available.

As the monkey_patch can contain code from other parties, which does not conform
to module_sync's rubocopy styles, this file should is ignored explicitly in
rubocops configuration.