Never use .erb in the config, prefer it in filenames
#93 changed modulesync to allow for .erb suffixed templates. In many cases
this is useful to avoid the template being interpreted as the real thing
by other tooling, like e.g. a `.gitattributes` file.

As discovered around #113, the original fix was not complete, and e.g.
failed when being combined with `:global` sections. This commit now moves
more towards always using `.erb` suffixes on template's names, and - for
readabilitie's sake - only look for un-adorned keys in the config files.

Closes #113