(maint) Preserve original line endings in a file
Previously when rendering a file the line endings would be converted by ruby
to the OS specific style e.g. CRLF on Windows.  However as many modules are
cross platform it is important to preserve the line ending style from the
original moduleroot file e.g. a .BAT file is for windows and should always use
CRLF, whether it's on a Linux of Windows system.  This commit changes the file
opening mode to use binary (which stops the line ending conversion) and forces
all file encoding to UTF-8 which is the default already.