Add GitLab merge request (MR) support
Adds the capability of creating GitLab merge requests
following the approach currently implemented for GitHub.
We try to make the user experience simply by sticking to calling
the feature "PR" / "--pr" also for GitLab. Which of both options,
GitHub or GitLab, is chosen for processing the user request is
determined by the environment variable being set:
- GITHUB_TOKEN ... a PR on GitHub will be created
- GITLAB_TOKEN ... a MR on GitLab will be created