Tests: Implement a Puppet module remote repository faker
This commit provides a new helper class Faker::PuppetModuleRemoteRepo to ease the write of tests when tested operations should (or not) operate on the git repository of a puppet module.

This class provides following features:
 - Fake a remote repository by creating a bare initialization usable using a local remote url (ie. file://)
 - Create basic operations behind the scene using a temporary repo (e.g. write, add and push a file; read a file from a specified branch)
 - Populate the fake remote repo using a basic metadata.json
 - Simulate a read-only repo
 - Change the default branch
 - Count commits between two commit refs
 - Count commits made by an author on all the repo, or a specified branch

Note: All operations are made using "Faker <faker@example.com>" as author and committer.

This commit comes with a tiny Faker module to hold the config (ie. the working directory)