Rework PR/MR related code (1/2)
This commit adds a check to ensure no PR/MR are made with the same source
and target branch.

This commit implements a facade class (ie. GitService::Base) to prevent
adding duplicate code in the git services and remove any global options[] usage.

It lets SourceCode#open_pull_request to craft the right parameters to
pass to the GitService once, so it deduplicates the code in specialised
service.

By design, this allows to honor easily the default repo branch.