Disable Style/FetchEnvVar cop
This cop doesn't make a lot of sense. While with a regular has there is
a difference between .fetch() and [] (consider x['key'] || default where
x['key'] is false), this is not possible for ENV where all values are
either strings or nil. All cases where ENV[] is used in this gem are
properly checked prior to being used.